wpt_babel/content-security-policy/inheritance/iframe-all-local-schemes-inherit-self.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js">
</script>
<script src="/resources/testharnessreport.js">
</script>
<meta content="img-src 'self'" http-equiv="Content-Security-Policy"/>
<body>
 <script>
  "use strict";

function wait_for_error_from_frame(frame, test) {
  window.addEventListener('message', test.step_func(function (e) {
    if (e.source != frame.contentWindow) return;
    assert_equals(e.data, "load");
    frame.remove();
    test.done();
  }));
}
async_test(function (t) {
  var i = document.createElement('iframe');
  document.body.appendChild(i);
  var img = document.createElement('img');
  img.onload = t.step_func_done(function (_) {
    return i.remove();
  });
  img.onerror = t.unreached_func();
  i.contentDocument.body.appendChild(img);
  img.src = "{{location[server]}}/images/red-16x16.png";
}, "<iframe>'s about:blank inherits policy.");
async_test(function (t) {
  var i = document.createElement('iframe');
  i.srcdoc = "\n      <img src='{{location[server]}}/images/red-16x16.png'\n        onload='window.top.postMessage(\"load\", \"*\");'\n        onerror='window.top.postMessage(\"error\", \"*\");'\n      >\n    ";
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe srcdoc>'s inherits policy.");
async_test(function (t) {
  var i = document.createElement('iframe');
  var b = new Blob(["\n        <img src='{{location[server]}}/images/red-16x16.png'\n          onload='window.top.postMessage(\"load\", \"*\");'\n          onerror='window.top.postMessage(\"error\", \"*\");'\n        >\n      "], {
    type: "text/html"
  });
  i.src = URL.createObjectURL(b);
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe src='blob:...'>'s inherits policy.");
async_test(function (t) {
  var i = document.createElement('iframe');
  i.src = "data:text/html,<img src='{{location[server]}}/images/red-16x16.png'\n      onload='window.top.postMessage(\"load\", \"*\");'\n      onerror='window.top.postMessage(\"error\", \"*\");'\n    >";
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe src='data:...'>'s inherits policy.");
async_test(function (t) {
  var i = document.createElement('iframe');
  i.src = "javascript:\"<img src='{{location[server]}}/images/red-16x16.png'\n      onload='window.top.postMessage(\\\"load\\\", \\\"*\\\");'\n      onerror='window.top.postMessage(\\\"error\\\", \\\"*\\\");'\n    >\"";
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe src='javascript:...'>'s inherits policy.");
async_test(function (t) {
  var i = document.createElement('iframe');
  var b = new Blob(["\n        <img src='{{location[server]}}/images/red-16x16.png'\n          onload='window.top.postMessage(\"load\", \"*\");'\n          onerror='window.top.postMessage(\"error\", \"*\");'\n        >\n      "], {
    type: "text/html"
  });
  i.src = URL.createObjectURL(b);
  i.sandbox = 'allow-scripts';
  wait_for_error_from_frame(i, t);
  document.body.appendChild(i);
}, "<iframe sandbox src='blob:...'>'s inherits policy. (opaque origin sandbox)");
 </script>
</body>
